+++ src/lethalhabit/Tickable.java
@@ -1,5 +1,5 @@
 package lethalhabit;
 
 public interface Tickable {
-    void tick(float timeDelta);
+    void tick(Double timeDelta);
 }
